requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
django-tinymce==3.2.0
djangorestframework==3.12.2
Markdown==3.3.3
Pillow==8.1.1
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
